+++ src/main/java/org/bytedeco/javacpp/tools/InfoMap.java
@@ -44,7 +44,8 @@ public class InfoMap extends HashMap<String,List<Info>> {
         .put(new Info("basic/containers").cppTypes("std::bitset", "std::deque", "std::list", "std::map", "std::queue", "std::set",
                                                    "std::stack", "std::vector", "std::valarray", "std::pair", "std::forward_list",
                                                    "std::priority_queue", "std::unordered_map", "std::unordered_set"))
-        .put(new Info("basic/types").cppTypes("signed", "unsigned", "char", "short", "int", "long", "bool", "float", "double"))
+        .put(new Info("basic/types").cppTypes("signed", "unsigned", "char", "short", "int", "long", "bool", "float", "double",
+                                              "_Bool", "_Complex", "_Imaginary", "complex", "imaginary"))
 
         .put(new Info("__COUNTER__").cppText("#define __COUNTER__ 0"))
         .put(new Info(" __attribute__", "__declspec").annotations().skip())
@@ -82,9 +83,11 @@ public class InfoMap extends HashMap<String,List<Info>> {
         .put(new Info("float", "jfloat").valueTypes("float").pointerTypes("FloatPointer", "FloatBuffer", "float[]"))
         .put(new Info("double", "jdouble").valueTypes("double").pointerTypes("DoublePointer", "DoubleBuffer", "double[]"))
         .put(new Info("long double").cast().valueTypes("double").pointerTypes("Pointer"))
-        .put(new Info("std::complex<float>").cast().pointerTypes("FloatPointer", "FloatBuffer", "float[]"))
-        .put(new Info("std::complex<double>").cast().pointerTypes("DoublePointer", "DoubleBuffer", "double[]"))
-        .put(new Info("bool", "jboolean").cast().valueTypes("boolean").pointerTypes("BoolPointer", "boolean[]"))
+        .put(new Info("std::complex<float>", "float _Complex", "float _Imaginary", "float complex", "float imaginary").cast()
+            .pointerTypes("FloatPointer", "FloatBuffer", "float[]"))
+        .put(new Info("std::complex<double>", "double _Complex", "double _Imaginary", "double complex", "double imaginary").cast()
+            .pointerTypes("DoublePointer", "DoubleBuffer", "double[]"))
+        .put(new Info("_Bool", "bool", "jboolean").cast().valueTypes("boolean").pointerTypes("BoolPointer", "boolean[]"))
         .put(new Info("wchar_t", "WCHAR").cast().valueTypes("char").pointerTypes("CharPointer"))
         .put(new Info("const char").valueTypes("byte").pointerTypes("@Cast(\"const char*\") BytePointer", "String"))
         .put(new Info("boost::shared_ptr", "std::shared_ptr").annotations("@SharedPtr"))